requirements.txt
streamlit
python-dotenv
pypdf
openai
